add the complete bool flags and completed timestamps
if step 1 was already completed, then there is no need to fetch the secret
questions and do that part of the dance again. simply re-send the email and tell
the user that they have already completed step 1.

a user repeating step 2 is superflous and unhandled. the number of times a user
visits the 2nd step verification link will make absolutely no difference.
visiting once completes the authentication